requirements.txt
Django==5.1.6
environs==14.1.1
django-phonenumber-field==8.0.0
phonenumbers==8.13.55